extract generic http methods to helper class
