+++ src/oura_py/oura_client.py
@@ -1,13 +1,8 @@
-import requests
-from urllib3.exceptions import InsecureRequestWarning
-from typing import Dict
-from json import JSONDecodeError
 from datetime import date, timedelta
 import logging
-from .auth import PersonalTokenRequestHandler
+from .helpers import RequestManager
 from .exceptions import OuraPyException
 from .models import (
-    Result,
     PersonalInfo,
     RingConfig,
     SleepSummary,
@@ -39,37 +34,14 @@ class OuraClient:
         """
         self.url = f"https://{hostname}/{ver}/usercollection"
         self._personal_access_token = personal_access_token
-        self._handler = PersonalTokenRequestHandler(personal_access_token)
         self._ssl_verify = ssl_verify
         self._logger = logger or logging.getLogger(__name__)
-        if not ssl_verify:
-            requests.packages.urllib3.disable_warnings(category=InsecureRequestWarning)
-
-    def get(self, endpoint: str, params: Dict = None) -> Result:
-        """Sends a GET request to the specified endpoint with optional parameters.
-
-        Args:
-            endpoint (str): The API endpoint to send the GET request to.
-            params (Dict, optional): A dictionary of query parameters to include in the request. Defaults to None.
-
-        Returns:
-            Result: The result of the GET request.
-        """
-        return self._request(method="GET", endpoint=endpoint, params=params)
-
-    def post(self, endpoint: str, params: Dict = None, data: Dict = None) -> Result:
-        """
-        Sends a POST request to the specified endpoint with the given parameters and data.
-
-        Args:
-            endpoint (str): The API endpoint to send the request to.
-            params (Dict, optional): The query parameters to include in the request. Defaults to None.
-            data (Dict, optional): The data to include in the body of the request. Defaults to None.
-
-        Returns:
-            Result: The result of the POST request.
-        """
-        return self._request(method="POST", endpoint=endpoint, params=params, data=data)
+        self._manager = RequestManager(
+            personal_access_token=self._personal_access_token,
+            url=self.url,
+            ssl_verify=self._ssl_verify,
+            logger=self._logger,
+        )
 
     def get_sleep_summary(
         self,
@@ -116,6 +88,21 @@ class OuraClient:
             next_token=next_token,
         )
 
+    def get_personal_info(self) -> PersonalInfo:
+        result = self._manager.get("personal_info")
+        data = PersonalInfo(**result.data)
+        return data
+
+    def get_ring_config(self, document_id: str | None = None) -> RingConfig:
+        endpoint = (
+            "ring_configuration"
+            if document_id is None
+            else f"ring_configuration/{document_id}"
+        )
+        result = self._manager.get(endpoint=endpoint)
+        data = RingConfig(**result.data)
+        return data
+
     def _get_summary_generic(
         self,
         summary_type: str,
@@ -127,30 +114,15 @@ class OuraClient:
     ):
         if next_token:
             self._logger.debug(msg=f"next_token={next_token}")
-            result = self.get(f"daily_{summary_type}/{next_token}")
+            result = self._manager.get(f"daily_{summary_type}/{next_token}")
             data = data_class_datum(**result.data)
             return data
         start_date, end_date = self._prep_dates(start, end)
-        result = self.get(
+        result = self._manager.get(
             f"daily_{summary_type}",
             params={"start_date": start_date, "end_date": end_date},
         )
         data = data_class(**result.data)
-        return data
-
-    def get_personal_info(self) -> PersonalInfo:
-        result = self.get("personal_info")
-        data = PersonalInfo(**result.data)
-        return data
-
-    def get_ring_config(self, document_id: str | None = None) -> RingConfig:
-        endpoint = (
-            "ring_configuration"
-            if document_id is None
-            else f"ring_configuration/{document_id}"
-        )
-        result = self.get(endpoint=endpoint)
-        data = RingConfig(**result.data)
         return data
 
     def _prep_dates(
@@ -165,53 +137,3 @@ class OuraClient:
             self._logger.error(msg=log_msg)
             raise OuraPyException("Start date must be before end date.")
         return str(start), str(end)
-
-    def _request(
-        self, method: str, endpoint: str, params: Dict = None, data: Dict = None
-    ) -> Result:
-        """
-        Makes an HTTP request to the specified endpoint with the given method, parameters, and data.
-
-        Args:
-            method (str): The HTTP method to use for the request (e.g., 'GET', 'POST').
-            endpoint (str): The API endpoint to send the request to.
-            params (Dict, optional): The query parameters to include in the request. Defaults to None.
-            data (Dict, optional): The data to include in the request body. Defaults to None.
-
-        Returns:
-            Result: An object containing the status code, message, and data from the response.
-
-        Raises:
-            OuraPyException: If there is an error making the request or if the response contains bad JSON.
-        """
-        url = f"{self.url}/{endpoint}"
-        headers = {"Authorization": f"Bearer {self._personal_access_token}"}
-        log_pre = f"method={method}, url={url}, params={params}"
-        log_post = ", ".join(("success={}", "status_code={}", "message={}"))
-        try:
-            self._logger.debug(msg=log_pre)
-            response = requests.request(
-                method=method,
-                url=url,
-                headers=headers,
-                params=params,
-                data=data,
-                verify=self._ssl_verify,
-            )
-        except requests.exceptions.RequestException as e:
-            self._logger.error(msg=str(e))
-            raise OuraPyException("Error making request") from e
-        try:
-            data_out = response.json()
-        except (ValueError, JSONDecodeError) as e:
-            self._logger.error(msg=log_post.format(False, None, e))
-            raise OuraPyException("Bad JSON in response") from e
-        req_success = 299 >= response.status_code >= 200
-        log_line = log_post.format(req_success, response.status_code, response.reason)
-        if req_success:
-            self._logger.debug(msg=log_line)
-            return Result(
-                status_code=response.status_code, message=response.reason, data=data_out
-            )
-        self._logger.error(msg=log_line)
-        raise OuraPyException(f"{response.status_code}: {response.reason}")
